chore: Rename plugin files
